requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
gunicorn==20.1.0
google-api-python-client==2.27.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
langchain-openai
langchain-community
langchain-core
pypdf